accept formatter callback on prompt format function
