+++ packages/evals/src/evaluators/llm/answer-relevancy/prompts.ts
@@ -23,12 +23,22 @@ export function generateEvaluationStatementsPrompt({ output }: { output: string 
 export const REASON_TEMPLATE =
   'Explain the irrelevancy score where 0 is the lowest and {scale} is the highest for the LLM\'s response using this context:\n    Context:\n    Input: {input}\n    Output: {output}\n    Score: {score}\n    Outcomes: {outcomes}\n    \n    Rules:\n    - Explain score based on mix of direct answers and related context\n    - Consider both full and partial relevance\n    - Keep explanation concise and focused\n    - Use given score, don\'t recalculate\n    - Don\'t judge factual correctness\n    - Explain both relevant and irrelevant aspects\n    - For mixed responses, explain the balance\n      Format:\n      {\n          "reason": "The score is {score} because {explanation of overall relevance}"\n      }\n      Example Responses:\n      {\n          "reason": "The score is 7 because while the first statement directly answers the question, the additional context is only partially relevant"\n      }\n      {\n          "reason": "The score is 3 because while the answer discusses the right topic, it doesn\'t directly address the question"\n      }\n      ';
 
-export function generateReasonPrompt({ input, output, eval_result, settings, outcomes }: LLMEvaluatorReasonPromptArgs) {
-  return REASON_TEMPLATE.replace('{scale}', String(settings.scale))
-    .replace('{input}', input)
-    .replace('{output}', output)
-    .replace('{score}', String(eval_result.score))
-    .replace('{outcomes}', JSON.stringify(outcomes));
+export function generateReasonPrompt({
+  input,
+  output,
+  eval_result,
+  settings,
+  outcomes,
+  formatter,
+  template,
+}: LLMEvaluatorReasonPromptArgs) {
+  return formatter(template, {
+    input,
+    output,
+    eval_result: String(eval_result.score),
+    scale: String(settings.scale),
+    outcomes: JSON.stringify(outcomes),
+  });
 }
 
 export const EVAL_TEMPLATE =
@@ -36,12 +46,12 @@ export const EVAL_TEMPLATE =
 
 export function generateEvaluatePrompt({ input, statements }: { input: string; statements: string[] }) {
   return `Evaluate each statement's relevance to the input question, considering direct answers, related context, and uncertain cases.
-  
+
       Return JSON with array of outcome objects. Each outcome must include:
       - "outcome": "yes", "no", or "unsure"
       - "reason": Clear explanation of the outcome
       - "claim": The statement being evaluated
-  
+
       Outcome Guidelines:
       - "yes": Statement explicitly and directly answers the input question when it:
           * Contains specific answer to the question asked (e.g., "The color of the sky is blue")
@@ -49,7 +59,7 @@ export function generateEvaluatePrompt({ input, statements }: { input: string; s
           * Can stand alone as a complete answer
           * Contains appropriate question-type response (e.g., location for "where", person for "who")
           * Note: If statement is incorrect but directly addresses the question, mark as "unsure"
-  
+
       - "unsure": Statement shows partial relevance when it:
           * Discusses the type of information being asked about (e.g., mentions temperatures when asked about temperature)
           * Contains information about the answer without explicit statement
@@ -65,8 +75,8 @@ export function generateEvaluatePrompt({ input, statements }: { input: string; s
           * References locations or entities in the same category as what's being asked about
           * Provides relevant information without using explicit question-type terminology
           * Contains references to properties of the subject that relate to the question type
-  
-  
+
+
       - "no": Statement lacks meaningful connection to question when it:
           * Contains neither the subject nor the type of information being requested
           * Contains no terms related to what's being asked about
@@ -76,13 +86,13 @@ export function generateEvaluatePrompt({ input, statements }: { input: string; s
           * Discusses the subject but not the specific attribute being asked about
           * Note: Assessment is about connection to what's being asked, not factual accuracy
           * Contains no connection to what's being asked about (neither the subject nor the type of information requested)
-  
-      REMEMBER: 
+
+      REMEMBER:
       - If the statement contains words or phrases that are relevant to the input, it is partially relevant.
       - If the statement is a direct answer to the input, it is relevant.
       - If the statement is completely unrelated to the input or contains nothing, it is not relevant.
       - DO NOT MAKE A JUDGEMENT ON THE CORRECTNESS OF THE STATEMENT, JUST THE RELEVANCY.
-  
+
       STRICT RULES:
       - If a statement mentions the type of information being requested, it should be marked as "unsure" ONLY if it's discussing that type meaningfully (not just mentioning it)
       - Subject mentions alone are NOT enough for relevance - they must connect to what's being asked about
@@ -94,18 +104,18 @@ export function generateEvaluatePrompt({ input, statements }: { input: string; s
       - Measurement/quantity relevance counts as type-level relevance
       - Administrative/governance terms are only relevant if they relate to the question type
       - Descriptive facts about the subject should be marked as "no" unless they directly relate to the question type
-  
-  
+
+
       Examples of "no" statements:
           * "Japan has beautiful seasons" for "What is Japan's largest city?"
           * "Trees grow tall" for "How tall is Mount Everest?"
           * "The weather is nice" for "Who is the president?"
-  
+
       Example:
       Input: "What color is the sky during daytime?"
       Statements: [
         "The sky is blue during daytime",
-        "The sky is full of clouds", 
+        "The sky is full of clouds",
         "I had breakfast today",
         "Blue is a beautiful color",
         "Many birds fly in the sky",
@@ -158,28 +168,34 @@ export function generateEvaluatePrompt({ input, statements }: { input: string; s
               }}
           ]
       }}
-  
+
   The number of outcomes MUST MATCH the number of statements exactly.
-  
+
     Input:
     ${input}
-  
+
     Number of statements: ${statements.length === 0 ? '1' : statements.length}
-  
+
     Statements:
     ${statements}
-  
+
     JSON:
     `;
 }
 
-export async function generateEvaluationPrompt({ input, output, agent }: LLMEvaluatorEvalPromptArgs) {
-  const statementPrompt = generateEvaluationStatementsPrompt({ output });
-  const statements = await agent.generate(statementPrompt, {
-    output: z.object({
-      statements: z.array(z.string()),
-    }),
+export async function generateEvaluationPrompt({
+  input,
+  context,
+  settings,
+  output,
+  formatter,
+  template,
+}: LLMEvaluatorEvalPromptArgs) {
+  return formatter(template, {
+    input,
+    statements: [output]?.join(','),
+    ...settings,
+    context: context?.join(', ') || '',
+    output,
   });
-  const prompt = generateEvaluatePrompt({ input, statements: statements.object.statements });
-  return prompt;
 }
